VMwareEngineNetworkPeering: fix mappers from generated
