+++ pkg/controller/direct/vmwareengine/mapper.generated.go
@@ -20,78 +20,6 @@ import (
 	"github.com/GoogleCloudPlatform/k8s-config-connector/pkg/controller/direct"
 )
 
-func VMwareEngineNetworkPeeringObservedState_FromProto(mapCtx *direct.MapContext, in *pb.NetworkPeering) *krm.VMwareEngineNetworkPeeringObservedState {
-	if in == nil {
-		return nil
-	}
-	out := &krm.VMwareEngineNetworkPeeringObservedState{}
-	// MISSING: Name
-	out.CreateTime = direct.StringTimestamp_FromProto(mapCtx, in.GetCreateTime())
-	out.UpdateTime = direct.StringTimestamp_FromProto(mapCtx, in.GetUpdateTime())
-	out.State = direct.Enum_FromProto(mapCtx, in.GetState())
-	out.StateDetails = direct.LazyPtr(in.GetStateDetails())
-	// MISSING: PeerMtu
-	// MISSING: Uid
-	// (near miss): "Uid" vs "UID"
-	// MISSING: VmwareEngineNetwork
-	return out
-}
-func VMwareEngineNetworkPeeringObservedState_ToProto(mapCtx *direct.MapContext, in *krm.VMwareEngineNetworkPeeringObservedState) *pb.NetworkPeering {
-	if in == nil {
-		return nil
-	}
-	out := &pb.NetworkPeering{}
-	// MISSING: Name
-	out.CreateTime = direct.StringTimestamp_ToProto(mapCtx, in.CreateTime)
-	out.UpdateTime = direct.StringTimestamp_ToProto(mapCtx, in.UpdateTime)
-	out.State = direct.Enum_ToProto[pb.NetworkPeering_State](mapCtx, in.State)
-	out.StateDetails = direct.ValueOf(in.StateDetails)
-	// MISSING: PeerMtu
-	// MISSING: Uid
-	// (near miss): "Uid" vs "UID"
-	// MISSING: VmwareEngineNetwork
-	return out
-}
-func VMwareEngineNetworkPeeringSpec_FromProto(mapCtx *direct.MapContext, in *pb.NetworkPeering) *krm.VMwareEngineNetworkPeeringSpec {
-	if in == nil {
-		return nil
-	}
-	out := &krm.VMwareEngineNetworkPeeringSpec{}
-	// MISSING: Name
-	out.PeerNetwork = direct.LazyPtr(in.GetPeerNetwork())
-	out.ExportCustomRoutes = in.ExportCustomRoutes
-	out.ImportCustomRoutes = in.ImportCustomRoutes
-	out.ExchangeSubnetRoutes = in.ExchangeSubnetRoutes
-	out.ExportCustomRoutesWithPublicIP = in.ExportCustomRoutesWithPublicIp
-	out.ImportCustomRoutesWithPublicIP = in.ImportCustomRoutesWithPublicIp
-	// MISSING: PeerMtu
-	// (near miss): "PeerMtu" vs "PeerMTU"
-	out.PeerNetworkType = direct.Enum_FromProto(mapCtx, in.GetPeerNetworkType())
-	// MISSING: Uid
-	// MISSING: VmwareEngineNetwork
-	out.Description = direct.LazyPtr(in.GetDescription())
-	return out
-}
-func VMwareEngineNetworkPeeringSpec_ToProto(mapCtx *direct.MapContext, in *krm.VMwareEngineNetworkPeeringSpec) *pb.NetworkPeering {
-	if in == nil {
-		return nil
-	}
-	out := &pb.NetworkPeering{}
-	// MISSING: Name
-	out.PeerNetwork = VMwareEngineNetworkPeeringSpec_PeerNetwork_ToProto(mapCtx, in.PeerNetwork)
-	out.ExportCustomRoutes = in.ExportCustomRoutes
-	out.ImportCustomRoutes = in.ImportCustomRoutes
-	out.ExchangeSubnetRoutes = in.ExchangeSubnetRoutes
-	out.ExportCustomRoutesWithPublicIp = in.ExportCustomRoutesWithPublicIP
-	out.ImportCustomRoutesWithPublicIp = in.ImportCustomRoutesWithPublicIP
-	// MISSING: PeerMtu
-	// (near miss): "PeerMtu" vs "PeerMTU"
-	out.PeerNetworkType = direct.Enum_ToProto[pb.NetworkPeering_PeerNetworkType](mapCtx, in.PeerNetworkType)
-	// MISSING: Uid
-	// MISSING: VmwareEngineNetwork
-	out.Description = direct.ValueOf(in.Description)
-	return out
-}
 func VMwareEngineNetworkSpec_FromProto(mapCtx *direct.MapContext, in *pb.VmwareEngineNetwork) *krm.VMwareEngineNetworkSpec {
 	if in == nil {
 		return nil
